add: consolidate logic of claim into one component
